pyproject.toml
[project]
name = "cocktails-rag-mcp-server"
version = "0.1.0"
description = "Cocktails RAG MCP Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastmcp>=2.12.4",
    "jq>=1.10.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=1.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.1.1",
]


[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88